+++ __tests__/data_structures.ts
@@ -1,4 +1,4 @@
-import { InworldPacket } from '../src/entities/inworld_packet.entity';
+import { InworldPacket } from '../src/entities/packets/inworld_packet.entity';
 
 export interface RegenerateResponse {
   interactionId?: string;
