+++ __tests__/entities/emotion_behavior.entity.spec.ts
@@ -3,7 +3,7 @@ import { EmotionEvent } from '@proto/ai/inworld/packets/packets_pb';
 import {
   EmotionBehavior,
   EmotionBehaviorCode,
-} from '../../src/entities/emotion_behavior.entity';
+} from '../../src/entities/packets/emotion/emotion_behavior.entity';
 
 const mappingTestTable = [
   {
