+++ __tests__/entities/emotion_strength.entity.spec.ts
@@ -3,7 +3,7 @@ import { EmotionEvent } from '@proto/ai/inworld/packets/packets_pb';
 import {
   EmotionStrength,
   EmotionStrengthCode,
-} from '../../src/entities/emotion_strength.entity';
+} from '../../src/entities/packets/emotion/emotion_strength.entity';
 
 const mappingTestTable = [
   {
