+++ src/entities/packets/emotion/emotion_behavior.entity.ts
