+++ src/entities/packets/emotion/emotion_strength.entity.ts
