+++ src/services/feedback.service.ts
@@ -1,5 +1,5 @@
 import { DislikeType, Feedback } from '../entities/feedback.entity';
-import { InworldPacket } from '../entities/inworld_packet.entity';
+import { InworldPacket } from '../entities/packets/inworld_packet.entity';
 import { ConnectionService } from './connection.service';
 import { FeedbackClientGrpcService } from './gprc/feedback_client_grpc.service';
